feat(config): debug mode
Introduces `--debug` flag (and `cfg.DebugMode`) to enable debugging
features for development purposes.

`dev` overlay has `DEBUG_MODE` environment variable injected to enable
it for local development.

With this change Gin framework runs in `production` mode, rather than
`debug` (which is a default and manifests are WARNING in the logs). As a
consequence Gin logger is less verbose in production-like enviroment.

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>

; Conflicts:
;	maas-api/cmd/main.go
;	maas-api/internal/config/config.go

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>