chore: Remove unneeded ca-certificates package from Dockerfile
The core functionalities of the maas-api (listing models, creating tokens, supporting authN/authZ) are working properly without the `ca-certificates` package. Thus, removing this package.

The main motivation for the removal is compatibility with hermetic Konflux builds which were failing because of the presence of this package. Secondarily, it is always better to not install unneeded packages keeping the image small, and reducing attack surface (if any).