refactor: unify scripts under single directory
Move files under `deployment/scripts/*` to `scripts/`.

The main motivation of this, is because the `deployments` directory
contains the Kustomize manifests for deploying MaaS. This is the
directory that the ODH operator would bundle in its image. Thus, these
files are extracted so that ODH operator won't include any executable
code.

Signed-off-by: Edgar Hernández <ehernand@redhat.com>