build(operator): add overlay for ODH
This adds a new `odh` Kustomize overlay. This is going to be the overlay that ODH operator will use for MaaS setup. The `openshift` overlay is adapted to be based on `odh` overlay, because the `openshift` one covers more (or said in another way, which could be more meaningful: the new odh overlay is covering less).

Despite the changes in the `openshift` overlay, the generated manifest with `kustomize build `deployment/overlays/openshift` is identical to the one generated without these changes.

When comparing the `odh` overlay with the `openshift` one:

* The `odh` overlay won't install the Gateway, because end-users are expected to create it prior to installing ODH.
* The `odh` overlay won't install RateLimitPolicy nor TokenRateLimitPolicy resources, because there may be no good defaults generally suitable. Users should bring their own policies as post-setup.
* The `odh` overlay won't install observability related resources, because no observability features are planned for the initial ODH-integrated release.

Essentially, the `openshift` overlay is becoming a developer overlay to setup MaaS in a more complete way for testing and development, while the `odh` overlay becomes the production setup.